Made cleanup not terrible; fixed a stupid ass-bug; now with 110% more --add-torrent!
I still can't tell if it's deleting guid folders from unrar or not
because it just smiles and says HOORAY YOUR SHIT'S DONE even when it's
not. Getting the video file was actually looping through every single
file even after it did the unrar, so it would end up picking anything
that was larger than any of the rars, like for example a goddamned
Sample. Seriously, why the fuck do those still exist? Does ANYONE use
them? Goddammit. And added the --add-torrent and --type options for
manually adding torrents since I'm way too lazy to actually make the
selenium part of this thing.