I'm probably going in the wrong direction with this. Added TZ offset for timestamps.
If a new episode gets added before the show is actually over, or at
least before the timestamp whose timezone is totally wrong + duration
has passed, it'll get updated on the next search for schedule updates.
So we can either make sure the searches know better, or try to hit
exact time targets for the timestamps. This should be fun.